Move route components in a specific file
