+++ src/theme/components/Link.tsx
@@ -16,17 +16,10 @@
 import styled from 'styled-components';
 
 import { styled as MaterialStyled, Link as MaterialLink } from '@mui/material';
-import { Link as GatsbyLink } from 'gatsby';
 
 type LinkProps = {
   selected?: boolean;
   href?: string;
-  onClick?: () => void;
-};
-
-type RouteProps = {
-  selected?: boolean;
-  to?: string;
   onClick?: () => void;
 };
 
@@ -97,40 +90,6 @@ const LinkWithMaterial = MaterialStyled(MaterialLink)<LinkProps>(
 `,
 );
 
-const RouteWithMaterial = MaterialStyled(GatsbyLink)<RouteProps>(
-  ({ theme, selected, to, onClick }) => `
-  text-decoration: none;
-  position: relative;
-  margin-bottom: 0;
-  padding-bottom: 3px;
-  color: inherit;
-  font-weight: 600;
-  ${selected && `border-bottom: 3px solid ${theme.palette.primary.main}`};
-  transition: 0.4s;
-  cursor: ${onClick || to ? 'pointer' : 'default'};
-
-  &:after {
-    content: '';
-    position: absolute;
-    right: 0;
-    width: 0;
-    bottom: -3px;
-    background: ${theme.palette.secondary.main};
-    height: 3px;
-    transition-property: width;
-    transition-duration: 0.3s;
-    transition-timing-function: ease-out;
-  }
-
-  &:focus:after,
-  &:hover:after {
-    left: 0;
-    right: auto;
-    width: 100%;
-  }
-`,
-);
-
 const LinkInButton = styled.a`
   text-decoration: none;
   display: block;
@@ -139,4 +98,4 @@ const LinkInButton = styled.a`
   color: inherit;
 `;
 
-export { Link, LinkInButton, LinkWithMaterial, RouteWithMaterial };
+export { Link, LinkInButton, LinkWithMaterial };
